weibo.cn/downloader/requirements.txt
requests
bs4
Pillow